requirements.txt
Flask==3.0.3
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
Werkzeug==3.0.3
passlib==1.7.4
pandas==2.2.2
openpyxl==3.1.5
reportlab==4.2.2
